+++ core/client_side/site_service/backend_api.js
@@ -1,18 +1,16 @@
 
+require('isomorphic-fetch');
 
-const fetch = require('fetch-cookie')(require('node-fetch'));
+const ROOT_URL = 'https://polyhedron-backend.herokuapp.com';
 
-const API_URL = 'https://polyhedron-backend.herokuapp.com';
-
-const URL_MAP = {
+const RELATIVE_URL_MAP = {
     register: '/api/user/registration',
     login: '/api/user/login',
     logout: '/api/user/logout',
     getUser: '/api/user/getuser',
     update: '/api/user/update',
     leaders: '/api/user/leaders/',
-    isLoggedIn: '/api/user/islogin',
-    flush: '/api/user/flush'
+    isLoggedIn: '/api/user/islogin'
 };
 
 const METHOD_MAP = {
@@ -23,8 +21,8 @@ const METHOD_MAP = {
 
 export class BackendAPI {
     constructor(rootURL, urlMap) {
-        this._rootURL = rootURL || API_URL;
-        this._urlMap = urlMap || URL_MAP;
+        this._rootURL = rootURL || ROOT_URL;
+        this._urlMap = urlMap || RELATIVE_URL_MAP;
     }
 
     register(email, login, password) {
@@ -58,17 +56,10 @@ export class BackendAPI {
         return this._fetchCORS(this._urlMap.isLoggedIn, METHOD_MAP.POST);
     }
 
-    flush() {
-        return this._fetchCORS(this._urlMap.flush, METHOD_MAP.POST);
-    }
-
-
     _fetchCORS(url, method, requestBody) {
         const absURL = this._rootURL + url;
 
-
         let options = {
-            body: requestBody ? JSON.stringify(requestBody) : '',
             method: method,
             mode: 'cors',
             credentials: 'include',
@@ -76,6 +67,9 @@ export class BackendAPI {
                 'Content-type': 'application/json'
             }
         };
+        if (method === METHOD_MAP.POST) {
+            options.body = requestBody ? JSON.stringify(requestBody) : '';
+        }
 
         return fetch(absURL,options);
     }
