+++ package.json
@@ -38,6 +38,8 @@
     "webpack": "^2.2.1"
   },
   "dependencies": {
-    "babel-runtime": "^6.23.0"
+    "babel-runtime": "^6.23.0",
+    "es6-promise": "^4.1.0",
+    "isomorphic-fetch": "^2.2.1"
   }
 }
