+++ static/js/pages/error404.js
@@ -1,6 +1,7 @@
 'use strict';
 import {BasePage} from './base';
 
+
 export class Error extends BasePage {
     render () {
         this._heading.innerHTML = "Ошибка";
