+++ templates/render_top.pug
@@ -3,7 +3,7 @@ if window.user
     |
     span.divider-dot &#183;
     |
-    a(data-page="leaders" href="#") Привет, #{user.name}!
+    a(data-page="leaders" href="#") Привет, #{user.login}!
     |
     span.divider-dot &#183;
     |
